% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.coastline.R
\name{assemble.coastline}
\alias{assemble.coastline}
\title{Assemble a coastline segment with repeating MPAs}
\usage{
assemble.coastline(Fr, Rw)
}
\arguments{
\item{Fr}{fraction of coastline in reserve}

\item{Rw}{reserve width (relative to dispersal distance)}
}
\description{
Assemble a coastline segment with repeating MPAs
}
