% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.matrix.R
\name{dispersal.matrix}
\alias{dispersal.matrix}
\title{Create connectivity matrix for MPA persistence calculations}
\usage{
dispersal.matrix(X, Mu, Sig)
}
\arguments{
\item{X}{the coastline vector (this will define the size of the matrix)}

\item{Mu}{the mean dispersal kernel}

\item{Sig}{the SD of dispersal kernel}
}
\description{
Create connectivity matrix for MPA persistence calculations
}
