% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPA_persist.R
\name{MPA_persist}
\alias{MPA_persist}
\title{This function will calculate whether a given MPA configuration on an infinite coastline will persist (deterministically)}
\usage{
MPA_persist(Fr, Rw, CRT, FLEP.out)
}
\arguments{
\item{Fr}{fraction of coastline in MPA}

\item{Rw}{reserve width (relative to larval dispersal distance)}

\item{CRT}{critical replacement threshold}

\item{FLEP.out}{fraction of lifetime egg production in fished area (assume FLEP = 1 in MPA)}
}
\description{
This function will calculate whether a given MPA configuration on an infinite coastline will persist (deterministically)
}
