% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{persistence}
\alias{persistence}
\title{Determine whether a coastline with MPAs is persistent or not}
\usage{
persistence(X, D, CRT, FLEP.out)
}
\arguments{
\item{X}{landscape}

\item{D}{dispersal matrix}

\item{CRT}{minimum value of FLEP for replacement}

\item{FLEP.out}{value of FLEP in fished patches}
}
\description{
Determine whether a coastline with MPAs is persistent or not
}
